refactor: 공간 관리자의 맵 API 관련 로직을 `api/managerMap.ts`로 이동
